Talked to Mira; gonna refactor
The current, non-working, code attempts to use loops inside plates. I'm going to refactor to have a loop outside the plate to set up a tensor with all the distribution parameters for each precinct, then a plate with .to_event instead of a loop to sample from all those distributions.