1.0.11: OK I can't be in denial
I think the problem with 1.0.9 was actually that SDS_TO_REDUCE_BY was less than SDS_TO_SHRINK_BY, so estimate of sigma was too low to fit data, so model tries to do things that will increase sigma. Better to have estimate of sigma too high. Let's try this...